requirements.txt
webdavclient3
python-dotenv
markdown
odfpy

